requirements.txt
requests
pandas
tqdm
bs4
loguru
dateparser
pymupdf
python-dotenv
openai